+++ README.md
@@ -1,3 +1,3 @@
 # benji
 
-Collects meta data about the system (CPU info, current running processes, etc.) and periodically records it in JSON format to a configurable IP
+Collects meta data about the system (CPU info, current running processes, etc.) and periodically records it in JSON format
