+++ src/main.c
@@ -1,0 +1,5 @@
+#include "include/utils.h"
+
+int main(int argc, const char* argv[]) {
+    winsock_init();
+}
