--- src/windows/include/system_info.hpp
@@ -1,11 +1,0 @@
-#ifndef _BENJI_SYSTEM_INFO_HPP
-#define _BENJI_SYSTEM_INFO_HPP
-
-#include <string>
-#include <array>
-#include <intrin.h>
-#include <cstring>
-
-std::string get_cpu_name();
-
-#endif
